@@ -18,36 +18,35 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  build_runner: ^1.12.2
-  charts_flutter: ^0.9.0
+  flutter_localizations:
+    sdk: flutter
+  build_runner: ^2.0.1
+  fl_chart: ^0.36.0
   cupertino_icons: ^1.0.2
   device_preview: ^0.7.1
   equatable: ^2.0.0
   flutter_bloc: ^7.0.0
-  flutter_secure_storage: ^4.1.0
-  intl: ^0.16.0
-  json_serializable: ^4.1.0
+  flutter_secure_storage: ^4.2.0
+  intl: ^0.17.0
+  json_serializable: ^4.1.1
   provider: ^5.0.0
   shared_preferences: ^2.0.5
-  table_calendar: ^2.3.3
+  table_calendar: ^3.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-dependency_overrides:
-  intl: ^0.17.0-nullsafety.2
-
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter.
 flutter:
-
+  generate: true
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
