@@ -1,6 +1,4 @@
 import 'dart:convert';
-
-import 'package:flutter/foundation.dart';
 import 'package:http/http.dart';
 import 'package:novaone/api/api.dart';
 import 'package:novaone/models/models.dart';
@@ -9,7 +7,7 @@ import 'package:novaone/novaOneUrl.dart';
 class UserApiClient extends BaseApiClient {
   final Client client;
 
-  UserApiClient({@required this.client}) : super(client: client);
+  UserApiClient({required this.client}) : super(client: client);
 
   /// Gets the user object from the API by sending a [email] and [password]
   /// to the NovOne Api
@@ -17,7 +15,7 @@ class UserApiClient extends BaseApiClient {
   /// Returns an [ApiMessageException] object if the request fails
   /// and a [User] object if the request was successful
   Future<User> getUser(
-      {@required String email, @required String password}) async {
+      {required String email, required String password}) async {
     Map<String, String> parameters = {'email': email, 'password': password};
 
     final response = await postToNovaOneApi(
