@@ -1,6 +1,6 @@
 class LeadsPerMonth {
-  final String date;
-  final int numberOfleads;
+  final String? date;
+  final int? numberOfleads;
 
   LeadsPerMonth({this.date, this.numberOfleads});
 }
