@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 
 class User extends Equatable {
   final int id;
@@ -23,34 +22,20 @@ class User extends Equatable {
   }
 
   User(
-      {@required this.id,
-      @required this.userId,
-      @required this.password,
-      @required this.lastLogin,
-      @required this.username,
-      @required this.firstName,
-      @required this.lastName,
-      @required this.email,
-      @required this.dateJoined,
-      @required this.isPaying,
-      @required this.wantsSms,
-      @required this.wantsEmailNotifications,
-      @required this.phoneNumber,
-      @required this.customerType})
-      : assert(id != null),
-        assert(userId != null),
-        assert(password != null),
-        assert(lastLogin != null),
-        assert(username != null),
-        assert(firstName != null),
-        assert(lastName != null),
-        assert(email != null),
-        assert(dateJoined != null),
-        assert(isPaying != null),
-        assert(wantsSms != null),
-        assert(wantsEmailNotifications != null),
-        assert(phoneNumber != null),
-        assert(customerType != null);
+      {required this.id,
+      required this.userId,
+      required this.password,
+      required this.lastLogin,
+      required this.username,
+      required this.firstName,
+      required this.lastName,
+      required this.email,
+      required this.dateJoined,
+      required this.isPaying,
+      required this.wantsSms,
+      required this.wantsEmailNotifications,
+      required this.phoneNumber,
+      required this.customerType});
 
   @override
   List<Object> get props => [
@@ -69,8 +54,7 @@ class User extends Equatable {
         phoneNumber,
         customerType
       ];
-  factory User.fromJson({@required Map<String, dynamic> json}) {
-    assert(json != null);
+  factory User.fromJson({required Map<String, dynamic> json}) {
     return User(
       id: json['id'],
       userId: json['userId'],
