@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 class OrientationLayout extends StatelessWidget {
-  final Widget landscape;
-  final Widget portrait;
+  final Widget? landscape;
+  final Widget? portrait;
 
-  const OrientationLayout({Key key, this.landscape, this.portrait})
+  const OrientationLayout({Key? key, this.landscape, this.portrait})
       : super(key: key);
 
   @override
@@ -12,9 +12,9 @@ class OrientationLayout extends StatelessWidget {
     final Orientation orientation = MediaQuery.of(context).orientation;
 
     if (orientation == Orientation.landscape) {
-      return landscape ?? portrait;
+      return landscape ?? portrait!;
     }
 
-    return portrait;
+    return portrait!;
   }
 }
