@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:novaone/enums/deviceScreenType.dart';
 
 class SizingInformation {
-  final DeviceScreenType deviceScreenType;
-  final Size screenSize;
-  final Size localWidgetSize;
+  final DeviceScreenType? deviceScreenType;
+  final Size? screenSize;
+  final Size? localWidgetSize;
 
   SizingInformation(
       {this.deviceScreenType, this.screenSize, this.localWidgetSize});
