@@ -4,6 +4,7 @@ import 'package:novaone/constants.dart';
 import 'package:novaone/screens/login/bloc/login_bloc.dart';
 import 'package:novaone/widgets/widgets.dart';
 import 'package:novaone/extensions/extensions.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class LoginTabletPortrait extends StatefulWidget {
   @override
@@ -17,10 +18,10 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
   FocusNode _focusNodePassWord = new FocusNode();
 
   // The user's email
-  String email;
+  String? email;
 
   // The user's password
-  String password;
+  String? password;
 
   // Create a global key that uniquely identifies the Form widget
   // and allows validation of the form.
@@ -40,8 +41,8 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
             children: <Widget>[
               const SizedBox(height: 40),
               Text(
-                'Welcome',
-                style: Theme.of(context).textTheme.headline1.copyWith(
+                AppLocalizations.of(context)?.helloWorld ?? '',
+                style: Theme.of(context).textTheme.headline1!.copyWith(
                     fontSize: 32,
                     fontWeight: FontWeight.bold,
                     color: Colors.black),
@@ -51,7 +52,7 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
               ),
               Text(
                 'Sign in to continue!',
-                style: Theme.of(context).textTheme.headline3.copyWith(
+                style: Theme.of(context).textTheme.headline3!.copyWith(
                     color: Colors.grey[400],
                     fontWeight: FontWeight.bold,
                     fontSize: 24),
@@ -78,10 +79,16 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
                   },
                   controller: emailController,
                   autovalidateMode: AutovalidateMode.onUserInteraction,
-                  validator: (email) => email.isValidEmail() && email != null
-                      ? null
-                      : 'Please enter a valid email',
-                  border: Border.all(color: Colors.grey[300], width: 2),
+                  validator: (email) {
+                    if (email != null) {
+                      if (email.isValidEmail()) {
+                        return null;
+                      }
+                    }
+
+                    return 'Please enter a valid email';
+                  },
+                  border: Border.all(color: Colors.grey[300]!, width: 2),
                   keyboardType: TextInputType.emailAddress,
                   hintText: 'Your Email',
                   labelText: 'Email Address',
@@ -101,10 +108,16 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
                     });
                   },
                   autovalidateMode: AutovalidateMode.onUserInteraction,
-                  validator: (password) =>
-                      password.isEmpty ? 'Please enter your password' : null,
+                  validator: (password) {
+                    if (password != null) {
+                      return password.isEmpty
+                          ? 'Please enter a password'
+                          : null;
+                    }
+                    return 'Please enter a password';
+                  },
                   controller: passwordController,
-                  border: Border.all(color: Colors.grey[300], width: 2),
+                  border: Border.all(color: Colors.grey[300]!, width: 2),
                   keyboardType: TextInputType.visiblePassword,
                   hintText: 'Your Password',
                   obscureText: true,
@@ -119,7 +132,7 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
                 'Forgot your password?',
                 style: Theme.of(context)
                     .textTheme
-                    .bodyText2
+                    .bodyText2!
                     .copyWith(fontSize: 16, fontWeight: FontWeight.bold),
               ),
               const SizedBox(
@@ -128,7 +141,7 @@ class _LoginTabletPortraitState extends State<LoginTabletPortrait> {
               NovaOneButton(
                 onPressed: () {
                   // First check if fields are empty before submitting request to API
-                  if (_formKey.currentState.validate()) {
+                  if (_formKey.currentState!.validate()) {
                     BlocProvider.of<LoginBloc>(context).add(LoginButtonTapped(
                         email: emailController.text,
                         password: passwordController.text));
