@@ -8,10 +8,8 @@ class StatefulWrapper extends StatefulWidget {
   final Function onInit;
   final Widget child;
 
-  const StatefulWrapper({Key key, @required this.onInit, @required this.child})
-      : assert(onInit != null),
-        assert(child != null),
-        super(key: key);
+  const StatefulWrapper({Key? key, required this.onInit, required this.child})
+      : super(key: key);
   @override
   _StatefulWrapperState createState() => _StatefulWrapperState();
 }
@@ -19,9 +17,7 @@ class StatefulWrapper extends StatefulWidget {
 class _StatefulWrapperState extends State<StatefulWrapper> {
   @override
   void initState() {
-    if (widget.onInit != null) {
-      widget.onInit();
-    }
+    widget.onInit();
     super.initState();
   }
 
