@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:novaone/enums/deviceScreenType.dart';
 
 DeviceScreenType getDeviceScreenType(
-    {@required MediaQueryData mediaQueryData}) {
+    {required MediaQueryData mediaQueryData}) {
   double deviceWidth = mediaQueryData.size.shortestSide;
 
   if (deviceWidth > 950) {
