@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:meta/meta.dart';
 
 ///
 /// Helper class that ensures a Widget is visible when it has the focus
@@ -29,9 +28,9 @@ import 'package:meta/meta.dart';
 ///
 class EnsureVisibleWhenFocused extends StatefulWidget {
   const EnsureVisibleWhenFocused({
-    Key key,
-    @required this.child,
-    @required this.focusNode,
+    Key? key,
+    required this.child,
+    required this.focusNode,
     this.curve: Curves.ease,
     this.duration: const Duration(milliseconds: 100),
   }) : super(key: key);
@@ -66,12 +65,12 @@ class _EnsureVisibleWhenFocusedState extends State<EnsureVisibleWhenFocused>
   void initState() {
     super.initState();
     widget.focusNode.addListener(_ensureVisible);
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
   }
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     widget.focusNode.removeListener(_ensureVisible);
     super.dispose();
   }
@@ -121,8 +120,8 @@ class _EnsureVisibleWhenFocusedState extends State<EnsureVisibleWhenFocused>
     }
 
     // Find the object which has the focus
-    final RenderObject object = context.findRenderObject();
-    final RenderAbstractViewport viewport = RenderAbstractViewport.of(object);
+    final RenderObject? object = context.findRenderObject();
+    final RenderAbstractViewport? viewport = RenderAbstractViewport.of(object);
 
     // If we are not working in a Scrollable, skip this routine
     if (viewport == null) {
@@ -130,14 +129,13 @@ class _EnsureVisibleWhenFocusedState extends State<EnsureVisibleWhenFocused>
     }
 
     // Get the Scrollable state (in order to retrieve its offset)
-    ScrollableState scrollableState = Scrollable.of(context);
-    assert(scrollableState != null);
+    ScrollableState scrollableState = Scrollable.of(context)!;
 
     // Get its offset
     ScrollPosition position = scrollableState.position;
     double alignment;
 
-    if (position.pixels > viewport.getOffsetToReveal(object, 0.0).offset) {
+    if (position.pixels > viewport.getOffsetToReveal(object!, 0.0).offset) {
       // Move down to the top of the viewport
       alignment = 0.0;
     } else if (position.pixels <
