@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:novaone/constants.dart';
 
 class MaxWidthContainer extends StatelessWidget {
-  final Widget child;
+  final Widget? child;
 
-  const MaxWidthContainer({Key key, this.child}) : super(key: key);
+  const MaxWidthContainer({Key? key, this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
